--- apps/charterafrica/src/components/EventsFellowshipsAndGrants/index.js
@@ -1,3 +1,0 @@
-import EventsFellowshipsAndGrants from "./EventsFellowshipsAndGrants";
-
-export default EventsFellowshipsAndGrants;
