+++ apps/charterafrica/src/components/OpportunityPage/OpportunityPage.snap.js
@@ -1,6 +1,6 @@
 // Jest Snapshot v1, https://goo.gl/fbAQLP
 
-exports[`<EventsFellowshipsAndGrants /> renders unchanged 1`] = `
+exports[`<OpportunityPage /> renders unchanged 1`] = `
 <div>
   <div
     class="MuiBox-root css-156ejva"
