+++ apps/charterafrica/src/components/OpportunityPage/index.js
@@ -1,0 +1,3 @@
+import OpportunityPage from "./OpportunityPage";
+
+export default OpportunityPage;
